requirements.txt
agora-py
rdflib
SPARQLWrapper
requests
python-dateutil
pyld
rdflib-jsonld
jsonpath-ng
shortuuid
networkx
gunicorn